requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.5
certifi==2020.6.20
cffi==1.14.0
chainer==4.2.0
cloudpickle==1.0.0
cupy==4.2.0
cycler==0.10.0
cytoolz==0.10.1
dask==1.2.2
decorator==4.4.0
enum34==1.1.6
fastrlock==0.4
filelock==3.0.10
functools32==3.2.3.post2
future==0.18.2
futures==3.3.0
imageio==2.6.1
kiwisolver==1.1.0
matplotlib==2.2.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.2
neural-renderer==1.1.3
numpy==1.16.3
olefile==0.46
Pillow==6.0.0
protobuf==3.7.1
pycparser==2.20
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
scikit-image==0.14.2
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
subprocess32==3.5.3
toolz==0.9.0
torch==1.3.1
torchvision==0.4.2
tornado==5.1.1
tqdm==4.32.1
typing==3.6.6
typing-extensions==3.6.6
tensorboard
